school Трохи оптимізував функції виклику з основого модуля
